refactored the delete and update files and put the php code into the functions file.

+++ demo/mysql/functions.php
@@ -43,36 +43,42 @@
   }
 
   function UpdateTable() {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if(isset($_POST['submit'])) {
+      global $connection;
+      $username = $_POST['username'];
+      $password = $_POST['password'];
+      $id = $_POST['id'];
 
-    $query = "UPDATE users SET ";
-    $query .= "username = '$username', ";
-    $query .= "password = '$password' ";
-    $query .= "WHERE id = $id ";
+      $query = "UPDATE users SET ";
+      $query .= "username = '$username', ";
+      $query .= "password = '$password' ";
+      $query .= "WHERE id = $id ";
 
-    $result = mysqli_query($connection, $query);
-    if(!$result) {
-      die("Update Query FAILED" . mysqli_error($connection));
-    } else {
-      echo "Record Updated!";
+      $result = mysqli_query($connection, $query);
+      if(!$result) {
+        die("Update Query FAILED" . mysqli_error($connection));
+      } else {
+        echo "Record Updated!";
+      }
     }
   }
 
   function DeleteRows() {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if(isset($_POST['submit'])) {
+      global $connection;
+      $username = $_POST['username'];
+      $password = $_POST['password'];
+      $id = $_POST['id'];
 
-    $query = "DELETE FROM users ";
-    $query .= "WHERE id = $id ";
+      $query = "DELETE FROM users ";
+      $query .= "WHERE id = $id ";
 
-    $result = mysqli_query($connection, $query);
-    if(!$result) {
-      die("Update Query FAILED" . mysqli_error($connection));
+      $result = mysqli_query($connection, $query);
+      if(!$result) {
+        die("Update Query FAILED" . mysqli_error($connection));
+      } else {
+        echo "Record Deleted!";
+      }
     }
   }
 ?>
